Add folder + OOI API for the ISS
Focused on making a working model for our objects of interest. At the moment only pulls the ISS from the Open-Notify API. While it moves all the passes of the ISS over Ruston at the moment, it would be easy to make it take it from any variable location.

Also, the function currently outputs a list as well as some text. The text comes from print statements in the function meant for debugging purposes.