Various Database Adaptations
Replaced the "temp" value in the Weather object in the database to respect high and low in order to fasilitate the creation of daily forecasts.  Values named "high" and "low" respectively

Further edited the code in the DarkSkyAPI reader to create those daily forecasts.

Furthermore changed values in models, routes, and some templetes to reflect the new format for the table.